extract rest query config
